Separated Errors and Parser Change
I separated the errors to ArgDefinitionError and ArgParserError.

While doing this I notice that the parser was not handling negative
numerical values(interger or float) nor was it handling float values
stating with a decimal ".".

I added the appropriate test to check the parser adjustments. I also
added the tests to check for the newly redefined errors.