Implemented toggleable boolean flags Added toggleable boolean flags this toggleability is only enabled for double dash flags. Added the supporting test to test the feature
Simplified the boolean parsing function and removed an unreachable
branch. The removed branch, was removed because of the way zig
ArgIterators work...once a white space is seen it doesn't get push to
the array/container it's skipped then it goes to the next argument.

The boolean parser now permits for TRUE and FALSE all cap as an
actionable input.