refactor(parser): remove double lookup on `parseNumeric()`
The function was doing a double look up on the item's container. Once to
get the active type then again to get the pointer to the active type and
assign a value. That is no more a single lookup for the active pointer
then pattern match the active type and assign the value.

Reference issues: Closes #11